% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceldown.R
\name{acceldown}
\alias{acceldown}
\title{The \code{acceldown} function estimates the relative locations of a vehicle downstream of a bottleneck.}
\usage{
acceldown(tstart, tend, umn, usd, xstart,  step)
}
\arguments{
\item{tstart}{start time in seconds, a number}

\item{tend}{end time, a number}

\item{umn}{start speed (fps) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location of the vehicle (feet), a number}

\item{step}{size in seconds, a number}
}
\value{
\code{acceldown} uses a stochastic model to show the location of the vehicle after it merges at a bottleneck.
}
\description{
The \code{acceldown} function estimates the relative locations of a vehicle downstream of a bottleneck.
}
\examples{
acceldown(8.75, 33.25, 47.59, 16.133, -70.36, 0.25)
}
