% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfree.R
\name{bmfree}
\alias{bmfree}
\title{The \code{bmfree} function returns a \code{tux} matrix of time \code{t}, speed \code{u} and location \code{x} values.}
\usage{
bmfree(umn, usd, N, Tup)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{umn}, a number}

\item{N}{number of time-steps, a number}

\item{Tup}{upper time range in minutes, a number}
}
\description{
The \code{bmfree} function returns a \code{tux} matrix of time \code{t}, speed \code{u} and location \code{x} values.
}
