% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfree2.R
\name{bmfree2}
\alias{bmfree2}
\title{The \code{bmfree2} function returns a \code{tux} matrix of time \code{t}, speed \code{u} (fps)
 and location \code{x} (feet).}
\usage{
bmfree2(umn, usd, tstart, tend, xstart, step, type)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{usd} (mph), a number}

\item{tstart}{simulation start time, a number}

\item{tend}{simulation end time, a number}

\item{xstart}{location (feet), a number}

\item{step}{time-step size, a number}

\item{type}{logical for plotting where TRUE creates a plot}
}
\description{
The \code{bmfree2} function returns a \code{tux} matrix of time \code{t}, speed \code{u} (fps)
 and location \code{x} (feet).
}
