% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkcross0.R
\name{brkcross0}
\alias{brkcross0}
\title{\code{brkcross0} estimates the time, speed and location that a vehicle crosses bottleneck location \code{x} = 0.}
\usage{
brkcross0(vehicle, df)
}
\arguments{
\item{vehicle, }{a number}

\item{df, }{a dataframe}
}
\description{
\code{brkcross0} estimates the time, speed and location that a vehicle crosses bottleneck location \code{x} = 0.
}
