% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkdelay.R
\name{brkdelay}
\alias{brkdelay}
\title{\code{brkdelay} is a wrapper function for \code{brktrials2}.}
\usage{
brkdelay(niter, tend, umn, usd, xstart, xfunnel, leff, lane, step)
}
\arguments{
\item{niter}{number of iterations, a number}

\item{tend}{end time for a simulation run, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{a vector of  \code{nveh} start locations, (feet), a vector}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle length(feet), a number}

\item{lane}{a vector of \code{nveh} numbers, a vector}

\item{step}{size in seconds, a number}
}
\value{
\code{brkdelay} returns  \code{t-x} trajectories of \code{nveh} vehicles at a bottleneck.
}
\description{
\code{brkdelay} is a wrapper function for \code{brktrials2}.
}
