% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brkdelay3.R
\name{brkdelay3}
\alias{brkdelay3}
\title{\code{brkdelay3} is a wrapper function for \code{brktrials3}.}
\usage{
brkdelay3(niter, nveh, tstart, tend, umn, usd, xstart, xfunnel, leff, step)
}
\arguments{
\item{niter}{number of iterations, a number}

\item{nveh}{number of vehicles, a number}

\item{tstart}{start time for a simulation run, a number}

\item{tend}{end time for a simulation run, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location, (feet), a number}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle length(feet), a number}

\item{step}{size in seconds, a number}
}
\value{
\code{brkdelay3} returns  \code{t-x} trajectories of \code{nveh} vehicles at a bottleneck.
}
\description{
\code{brkdelay3} is a wrapper function for \code{brktrials3}.
}
