% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brksummary.R
\name{brksummary}
\alias{brksummary}
\title{The function \code{brksummary} is a wrapper for  \code{accelpass} and \code{passplot}.}
\usage{
brksummary(k, tend, umn, usd, xstart, xfunnel, leff, lane, step)
}
\arguments{
\item{k}{traffic density (vehicles per mile, vpm), a number}

\item{tend}{end time, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location for vehicle in lane 1 (feet), a number}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle length(feet), a number}

\item{lane}{number, a number}

\item{step}{size in seconds, a number}
}
\value{
\code{brksummary} contains a table of information used for \code{run} plot
}
\description{
The function \code{brksummary} is a wrapper for  \code{accelpass} and \code{passplot}.
}
\examples{
brksummary(50, 30, 41, 11, -1000, -500, 14, 0, 2)
}
