% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacityplot.R
\name{capacityplot}
\alias{capacityplot}
\title{The function \code{capacityplot} creates a \code{(k,Q*)} plot indicating the success rate of sustaining capacity \code{Q*}}
\usage{
capacityplot(QUKtdayX)
}
\arguments{
\item{QUKtdayX}{a data frame}
}
\description{
The function \code{capacityplot} creates a \code{(k,Q*)} plot indicating the success rate of sustaining capacity \code{Q*}
}
\examples{
\donttest{capacityplot(QUKtdayX)}
}
