% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfamin.R
\name{cfanim}
\alias{cfanim}
\title{\code{cfanim} uses a stochastic model of speed to describe driver's ability to maintain a speed \code{u}.}
\usage{
cfanim(umn, usd, tup, dt)
}
\arguments{
\item{umn}{mean speed (mph), a number}

\item{usd}{standard deviation of \code{umn}, a number}

\item{tup}{upper time range, a number}

\item{dt}{time-step, a number}
}
\value{
The \code{cfanim} function returns information for a single vehicle
used to analyze the effects of traffic noise on car-following.
The function uses a stochastic traffic noise model, a Brownian motion or Wiener \code{W} model.
}
\description{
\code{cfanim} uses a stochastic model of speed to describe driver's ability to maintain a speed \code{u}.
}
