% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossconflict.R
\name{crossconflict}
\alias{crossconflict}
\title{The function \code{crossconflict} a graph showing driver conflict of two vehicles changing lanes.}
\usage{
crossconflict()
}
\description{
The function \code{crossconflict} a graph showing driver conflict of two vehicles changing lanes.
}
\examples{
crossconflict()
}
