% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{Space headway \code{h} between the lead and following vehicles at time \code{t}. Lead and following vehicles use \code{gbm} and constant speed models, respectively.}
\usage{
f1(xl0, ul0, theta, xf1, t, t1)
}
\arguments{
\item{xl0}{location of lead vehicle at \code{t} = 0, a number}

\item{ul0}{pre-breakdown speed of lead vehicle, a number}

\item{theta}{a \code{gbm} model parameter, a number}

\item{xf1}{location of following vehicle at \code{t} = \code{t1}, a number}

\item{t}{time, a number}

\item{t1}{deceleration time, a number}
}
\description{
Space headway \code{h} between the lead and following vehicles at time \code{t}. Lead and following vehicles use \code{gbm} and constant speed models, respectively.
}
