% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.R
\name{f2}
\alias{f2}
\title{Headway \code{h} between the lead and following vehicles at time \code{t}. Lead and following vehicles use constant speed models, respectively.}
\usage{
f2(xl0, ul0, ul3, theta, t3, xf1, t, t1)
}
\arguments{
\item{xl0}{location of lead vehicle at time \code{t} = \code{t0}, a number}

\item{ul0}{pre-breakdown speed of lead vehicle, a number}

\item{ul3}{breakdown speed of lead vehicle, a number}

\item{theta}{a \code{gbm} model parameter, a number}

\item{t3}{transition time, a number}

\item{xf1}{location of following vehicle at time \code{t} = \code{t1}, a number}

\item{t}{time, a number}

\item{t1}{deceleration time, a number}
}
\description{
Headway \code{h} between the lead and following vehicles at time \code{t}. Lead and following vehicles use constant speed models, respectively.
}
