% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findt1.R
\name{findt1}
\alias{findt1}
\title{\code{findt1} returns the deceleration time \code{t1}, the location \code{xf1}, and speed \code{uf1}
of the following vehicle at time \code{t1}.}
\usage{
findt1(xl0, xf0, ul0, uf0, theta, leff)
}
\arguments{
\item{xl0}{lead vehicle location at time \code{t0}, a number.}

\item{xf0}{following vehicle location at time \code{t0}, a number.}

\item{ul0}{lead vehicle speed, a number.}

\item{uf0}{following vehicle speed, a number.}

\item{theta}{a \code{gbm} model parameter, a number.}

\item{leff}{effective vehicle length, a number.}
}
\description{
\code{findt1} returns the deceleration time \code{t1}, the location \code{xf1}, and speed \code{uf1}
of the following vehicle at time \code{t1}.
}
