% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findt1ab.R
\name{findt1ab}
\alias{findt1ab}
\title{The \code{findt1ab} returns the time \code{tf1} and the location \code{xf1} and speed \code{uf1}
of the following vehicle when the driver begins to decelerate.}
\usage{
findt1ab(xl0, xf0, uf0, ul0, tseq, useq, t1, t2)
}
\arguments{
\item{xl0}{lead vehicle location at time \code{t0}, a number}

\item{xf0}{following vehicle location at time \code{t0}, a number}

\item{uf0}{following vehicle speed at time \code{t0}, a number}

\item{ul0}{lead vehicle speed at time \code{t0}, a number}

\item{tseq}{time sequence of the lead vehicle \code{l}, a vector}

\item{useq}{speed sequence of the lead vehicle \code{l}, a vector}

\item{t1}{start time, a number}

\item{t2}{end time, a number}
}
\description{
The \code{findt1ab} returns the time \code{tf1} and the location \code{xf1} and speed \code{uf1}
of the following vehicle when the driver begins to decelerate.
}
