% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findt2.R
\name{findt2}
\alias{findt2}
\title{\code{findt2} returns the time \code{t2} and \code{xf2} when the following vehicle driver recognizes
the lead vehicle has decelerated.}
\usage{
findt2(L, theta, t4)
}
\arguments{
\item{L}{lead vehicle information, a matrix}

\item{theta}{a \code{gbm} model parameter, a number}

\item{t4}{time, a number}
}
\description{
\code{findt2} returns the time \code{t2} and \code{xf2} when the following vehicle driver recognizes
the lead vehicle has decelerated.
}
