% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findt2ab.R
\name{findt2ab}
\alias{findt2ab}
\title{\code{t2} is the intersection point of the following driver sight-line and lead vehicle \code{x} trajectory.
Lead vehicle location \code{xl2} and speed \code{ul2} at the intersection point, respectively.}
\usage{
findt2ab(tstart, tmid, tend, ulstart, ulmid, xlstart, xlmid, xfstart, uf, lty, lwd, col)
}
\arguments{
\item{tstart}{start time, a number}

\item{tmid}{intermediate time, a number}

\item{tend}{end time, a number}

\item{ulstart}{lead vehicle starting speed at \code{tstart}, a number}

\item{ulmid}{lead vehicle end speed at \code{tmid}, a number}

\item{xlstart}{lead vehicle location at \code{tstart}, a number}

\item{xlmid}{lead vehicle location at \code{tmid}, a number}

\item{xfstart}{following vehicle location at \code{tstart}, a number}

\item{uf}{following vehicle speed, a number}

\item{lty}{line type, a number}

\item{lwd}{line width, a number}

\item{col}{line color, a factor}
}
\description{
\code{t2} is the intersection point of the following driver sight-line and lead vehicle \code{x} trajectory.
Lead vehicle location \code{xl2} and speed \code{ul2} at the intersection point, respectively.
}
