% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findt5.R
\name{findt5}
\alias{findt5}
\title{The time \code{t5} when the following vehicle reaches the bottleneck or location \code{x} = 0.}
\usage{
findt5(xf1, xf3, uf1, uf3, theta, t, t1, t2, t3, t4)
}
\arguments{
\item{xf1}{location of the following vehicle at \code{t1}, a number}

\item{xf3}{location of the following vehicle at \code{t3}, a number}

\item{uf1}{speed of the following vehicle at \code{t1}, a number}

\item{uf3}{speed of the following vehicle at \code{t3}, a number}

\item{theta}{a \code{gbm} model parameter of the following vehicle, a number}

\item{t}{unknown time \code{t5}, a number}

\item{t1}{deceleration time, a number}

\item{t2}{time when the vehicle sights lead vehicle, a number}

\item{t3}{time when the vehicle obtains breakdown speed \code{u0}, a number}

\item{t4}{upper end of search range, a number}
}
\description{
The time \code{t5} when the following vehicle reaches the bottleneck or location \code{x} = 0.
}
