% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findt5ab.R
\name{findt5ab}
\alias{findt5ab}
\title{The time \code{t5} when the following vehicle reaches the bottleneck or location \code{x} = 0.}
\usage{
findt5ab(xf1, xf2, uf1, uf2, t1, t2, t4)
}
\arguments{
\item{xf1}{location of the following vehicle at time \code{t1}, a number}

\item{xf2}{location of the following vehicle at time \code{t2}, a number}

\item{uf1}{speed of the following vehicle at time \code{t1}, a number}

\item{uf2}{speed of the following vehicle at time \code{t2}, a number}

\item{t1}{deceleration time, a number}

\item{t2}{time when the vehicle sights lead vehicle, a number}

\item{t4}{upper end of search range, a number}
}
\description{
The time \code{t5} when the following vehicle reaches the bottleneck or location \code{x} = 0.
}
