% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formqueue.R
\name{formqueue}
\alias{formqueue}
\title{The \code{formqueue} function is used to explain queuing upstream of a bottleneck.}
\usage{
formqueue(tstart, tend, umn, usd, xstart, xfunnel, leff)
}
\arguments{
\item{tstart}{start time, a number}

\item{tend}{end time, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location for vehicle in lane 1 (feet), a number}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle length(feet), a number}
}
\value{
\code{formqueue} uses a stochastic model to illustrate an ``idealistic'' situation,
a so-called a ``form queue merge.''
}
\description{
The \code{formqueue} function is used to explain queuing upstream of a bottleneck.
}
\examples{
formqueue(0, 10, 41, 11.6, -1000, -500, 14)
}
