% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmtheta.R
\name{gbmtheta}
\alias{gbmtheta}
\title{Estimate \code{theta} of a geometric Brownian motion model.}
\usage{
gbmtheta(upre, u0, t3)
}
\arguments{
\item{upre}{pre-breakdown speed, a number}

\item{u0}{breakdown speed, a number}

\item{t3}{time when u0 is initiated, a number}
}
\description{
Estimate \code{theta} of a geometric Brownian motion model.
}
