% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmx.R
\name{gbmx}
\alias{gbmx}
\title{Estimate distance \code{x} given a geometric Brownian motion model and time t.}
\usage{
gbmx(u0, theta, t)
}
\arguments{
\item{u0}{initial speed, a number}

\item{theta}{a \code{gbm} model estimate, a number.}

\item{t}{time, a number}
}
\description{
Estimate distance \code{x} given a geometric Brownian motion model and time t.
}
