% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h0.R
\docType{data}
\name{h0}
\alias{h0}
\title{Space headway \code{h0} at time \code{t} = 0.}
\format{\code{h0} is the minimum headway of 5280/\code{k0} and \code{hsafe(u0)} where \code{u0} is the lead vehicle speed at \code{t} = 0.
\describe{
\item{h0}{distance between lead and following vehicles, measured in feet.}
}}
\usage{
h0
}
\description{
Space headway \code{h0} at time \code{t} = 0.
}
\keyword{datasets}
