% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1.R
\docType{data}
\name{h1}
\alias{h1}
\title{Space headway \code{h1} at time \code{t1}, the time the following vehicle initiates deceleration.}
\format{\code{h1} is the minimum headway of based on the following vehicle sight-line and \code{hsafe(u1)} where \code{u1} is the lead vehicle speed at \code{t} = \code{t1}.
\describe{
\item{h1}{headway between lead and following vehicles, measured in feet.}
}}
\usage{
h1
}
\description{
Space headway \code{h1} at time \code{t1}, the time the following vehicle initiates deceleration.
}
\keyword{datasets}
