% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\docType{data}
\name{h2}
\alias{h2}
\title{Space headway \code{h2} at time \code{t2}, the time the following vehicle driver sees the lead vehicle has decelerated.}
\format{\code{h2} is the headway \code{hsafe(u2)} where \code{u2} is the lead vehicle speed at \code{t} = \code{t2}.
\describe{
\item{h2}{headway between lead and following vehicles, measured in feet.}
}}
\usage{
h2
}
\description{
Space headway \code{h2} at time \code{t2}, the time the following vehicle driver sees the lead vehicle has decelerated.
}
\keyword{datasets}
