% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3.R
\docType{data}
\name{h3}
\alias{h3}
\title{Space headway \code{h3} at time \code{t3}, the time the lead vehicle has decelerated to breakdown speed \code{u0}.}
\format{\code{h3} is the headway \code{hsafe(u3)} where \code{u3} = \code{u0}, the lead vehicle speed at \code{t} = \code{t3}.
\describe{
\item{h3}{headway between lead and following vehicles, measured in feet.}
}}
\usage{
h3
}
\description{
Space headway \code{h3} at time \code{t3}, the time the lead vehicle has decelerated to breakdown speed \code{u0}.
}
\keyword{datasets}
