\name{lane}
\alias{lane}
\docType{data}
\title{
\code{lane} is a vector that designates the lane drivers are traveling in the \code{brktrials} function.
}
\format{A vector used by \code{brktrials}
\describe{
\item{lane}{a lane number.}
}}
\description{ \code{lane} is a input vector to \code{brtrials}. For example, the lanes may be numbered: 0 1 2 1 2.}
\keyword{datasets}
