% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanedf.R
\name{lanedf}
\alias{lanedf}
\title{\code{lanedf} returns a vector of vehicles \code{veh} in \code{lane}.}
\usage{
lanedf(pick, nveh, lane)
}
\arguments{
\item{pick}{a lane number}

\item{nveh, }{a number}

\item{lane, }{a vector}
}
\description{
\code{lanedf} returns a vector of vehicles \code{veh} in \code{lane}.
}
