% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leff.R
\docType{data}
\name{leff}
\alias{leff}
\title{\code{leff} is the effective length of a vehicle.}
\format{\code{leff} is the length of a vehicle.
\describe{
\item{leff}{effective vehicle length, measured in feet.}
}}
\usage{
leff
}
\description{
\code{leff} is the effective length of a vehicle.
}
\keyword{datasets}
