% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfmatrix.R
\name{lfmatrix}
\alias{lfmatrix}
\title{The function \code{lfmatrix} displays \code{cflist}, the lists \code{L} and \code{Fw} information
for lead and following vehicles, respectively.}
\usage{
lfmatrix(cflist, nveh)
}
\arguments{
\item{cflist}{information produced from the \code{run} and \code{trial} functions, a list.}

\item{nveh}{number of vehicles in the investigation or \code{vehicle} = {1, 2, ..., \code{nveh}}, a number}
}
\value{
\code{cflist} returns two matrices consisting of speed \code{u} and location \code{x} data, which
are in turn are used to estimate performance measures for the simulation.
}
\description{
The function \code{lfmatrix} displays \code{cflist}, the lists \code{L} and \code{Fw} information
for lead and following vehicles, respectively.
}
