% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticmodel.R
\name{logisticmodel}
\alias{logisticmodel}
\title{The function \code{logisticmodel} estimates the parameters of logistic regression model
using a binomial formula and logit link function for data \code{QUKtdayX}.}
\usage{
logisticmodel(QUKtdayX, type)
}
\arguments{
\item{QUKtdayX}{a matrix}

\item{type}{logical}
}
\value{
If \code{type} is TRUE, plot a density-failure probability plot is created and a data frame is returned.
If If \code{type} is FALSE, a data frame is returned.
}
\description{
The function \code{logisticmodel} estimates the parameters of logistic regression model
using a binomial formula and logit link function for data \code{QUKtdayX}.
}
\examples{
\donttest{logisticmodel(QUKtdayX, TRUE)}
}
