% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge3.R
\name{merge3}
\alias{merge3}
\title{\code{merge3} produces \code{t-x} trajectories for a pair of lead and following vehicles for \code{brktrial3}.}
\usage{
merge3(i,df1,df2,leff,step,xfunnel,usd,ylim,type)
}
\arguments{
\item{i}{vehicle, a number}

\item{df1}{leading vehicle, a matrix}

\item{df2}{following vehicle, a matrix}

\item{leff}{vehicle length, a number}

\item{step}{time-step size, a number}

\item{xfunnel}{upstream location of bottleneck taper, a number}

\item{usd}{speed volatility (mph) for \code{umn}, a number}

\item{ylim}{for plot, a vector}

\item{type}{TRUE to create plots or FALSE otherwise, a logical}
}
\value{
The \code{merge3}, a wrapper function for \code{bmfree2}, \code{xabparam} and
\code{hsafe}, returns a smooth \code{hsafe} rule \code{t-x} trajectory.
for the following vehicle. The lead vehicle trajectory is not affected.
}
\description{
\code{merge3} produces \code{t-x} trajectories for a pair of lead and following vehicles for \code{brktrial3}.
}
