% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopass.R
\name{nopass}
\alias{nopass}
\title{\code{nopass} produces \code{t}, \code{u} and \code{x} for lead and following vehicles downstream of a bottleneck}
\usage{
nopass(veh, nope, leff)
}
\arguments{
\item{veh}{vehicle, a number}

\item{nope}{is a data.frame of leading and following vehicles: time, speed and location data, a data frame}

\item{leff}{effective vehicle length in feet, a number}
}
\value{
\code{brktrials} returns a data frame speed and location downstream of bottleneck.
}
\description{
\code{nopass} produces \code{t}, \code{u} and \code{x} for lead and following vehicles downstream of a bottleneck
}
