% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcrossover.R
\name{plotcrossover}
\alias{plotcrossover}
\title{A \code{ggplot2} plot of crossover traffic data.}
\usage{
plotcrossover(umn, leff)
}
\arguments{
\item{umn}{average speed in mph, a number}

\item{leff}{effective vehicle length in feet, a number}
}
\description{
A \code{ggplot2} plot of crossover traffic data.
}
\examples{
\donttest{plotcrossover(67, 14)}
}
