% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdesire3.R
\name{plotdesire3}
\alias{plotdesire3}
\title{\code{plotdesire3} creates \code{t-x} desire-line trajectories.}
\usage{
plotdesire3(tuxv, xfunnel)
}
\arguments{
\item{tuxv}{estimates from \code{brktrials3}, a matrix}

\item{xfunnel, }{location of the  vehicle, a vector}
}
\description{
\code{plotdesire3} creates \code{t-x} desire-line trajectories.
}
