% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfollow.R
\name{plotfollow}
\alias{plotfollow}
\title{Produces a time-distance \code{t-x} trajectory for \code{vehicle i = 2}.}
\usage{
plotfollow(L, Fw, t4)
}
\arguments{
\item{L}{a collection of variables that describe the relationship between a lead \code{l}
and a following \code{f} vehicle, a matrix.}

\item{Fw}{a collection of variables that describe the relationship between a lead \code{l}
and a following \code{f} vehicle, a matrix.}

\item{t4}{upper range of the plot, a number.}
}
\description{
Produces a time-distance \code{t-x} trajectory for \code{vehicle i = 2}.
}
