% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plowfollow3.R
\name{plotfollow3}
\alias{plotfollow3}
\title{\code{plotfollow3} adds a new trajectory to an existing graph.}
\usage{
plotfollow3(cflist, vehicle, uf, kf, ub, leff, t4)
}
\arguments{
\item{cflist}{contains information about the lead vehicle, a list}

\item{vehicle}{a vehicle identifier, a number}

\item{uf}{speed (mph) of the following vehicle at time \code{t0}, a number}

\item{kf}{density (vehicles per mile, vpm) of the following vehicle at time \code{t0}, a number}

\item{ub}{breakdown speed (mph) for this trajectory, a number}

\item{leff}{effective vehicle length, a number.}

\item{t4}{upper range of plot, a number}
}
\description{
\code{plotfollow3} adds a new trajectory to an existing graph.
}
