% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlead.R
\name{plotlead}
\alias{plotlead}
\title{Produces a time-distance \code{t-x} trajectory for the lead vehicle, \code{vehicle = 1}.}
\usage{
plotlead(L, theta, leff, xlim, ylim)
}
\arguments{
\item{L, }{a matrix}

\item{theta}{a \code{gbm} model parameter, a number}

\item{leff}{effective vehicle length (feet), a number}

\item{xlim}{plot range, a vector}

\item{ylim}{plot range, a vector}
}
\description{
Produces a time-distance \code{t-x} trajectory for the lead vehicle, \code{vehicle = 1}.
}
