% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmerge3.R
\name{plotmerge3}
\alias{plotmerge3}
\title{\code{plotmerge3} creates \code{t-x} safe headway trajectories.}
\usage{
plotmerge3(tuxv.fix, vehorder, xfunnel)
}
\arguments{
\item{tuxv.fix}{estimates from \code{brktrials3}, a matrix}

\item{vehorder}{vehicle arrivals at \code{x} = 0, a vector}

\item{xfunnel, }{location of the  vehicle, a vector}
}
\description{
\code{plotmerge3} creates \code{t-x} safe headway trajectories.
}
