% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotupstream.R
\name{plotupstream}
\alias{plotupstream}
\title{\code{plotupstream} produces \code{t-x} trajectories for lead and following vehicles at a bottleneck}
\usage{
plotupstream(pick, lane, nveh, df, xfunnel, leff, type)
}
\arguments{
\item{pick}{a lane number}

\item{lane, }{a vector}

\item{nveh, }{a number}

\item{df}{a data frame from \code{brktrials2}}

\item{xfunnel}{upstream location where the lane drop starts (feet), a number}

\item{leff}{effective vehicle length(feet), a number}

\item{type}{2 to create a plot, a number}
}
\value{
\code{plotupstream} returns \code{t-x} trajectories of \code{nveh} vehicles at a bottleneck.
Wrapper for \code{vehdf}, \code{nopass} and \code{vehfix}
}
\description{
\code{plotupstream} produces \code{t-x} trajectories for lead and following vehicles at a bottleneck
}
