\name{QUKtdayX}
\alias{QUKtdayX}
\docType{data}
\title{I-93 data set}
\description{A data frame with 19044 observations on the following 6 variables.}
\usage{data("QUKtdayX")}
\format{
  A data frame with 19044 observations on the following 6 variables.
  \describe{
    \item{\code{q}}{a numeric vector}
    \item{\code{u}}{a numeric vector}
    \item{\code{k}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{t}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
  }
}

