% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrheadway.R
\name{rrheadway}
\alias{rrheadway}
\title{\code{rrheadway} is a function to estimate the traffic flow \code{q} on a ring road.}
\usage{
rrheadway(LF, dt)
}
\arguments{
\item{LF}{location \code{x} output from \code{rrtrials}, a matrix.}

\item{dt}{time-step, a number.}
}
\value{
The \code{rrheadway} uses the output matrix \code{LF} from \code{rrtrials}.
}
\description{
\code{rrheadway} is a function to estimate the traffic flow \code{q} on a ring road.
}
