% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrheadwayu.R
\name{rrheadwayu}
\alias{rrheadwayu}
\title{\code{rrheadwayu} is a function used to find the speed \code{u} on a ring road at \code{x = 0}.}
\usage{
rrheadwayu(LFu, t5)
}
\arguments{
\item{LFu}{speed \code{u} output from \code{rrtrials}, a matrix.}

\item{t5}{time when vehicle crosses the line \code{x = 0}, a vector.}
}
\value{
The \code{rrheadwayu} uses the output matrix \code{LFu} from \code{rrtrials} and code{t5}.
}
\description{
\code{rrheadwayu} is a function used to find the speed \code{u} on a ring road at \code{x = 0}.
}
