% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safecrossover.R
\name{safecrossover}
\alias{safecrossover}
\title{The function \code{safecrossover} a graph showing driver conflict of two vehicles changing lanes.}
\usage{
safecrossover(umn, leff)
}
\arguments{
\item{umn}{average speed in mph, a number}

\item{leff}{effective vehicle length in feet, a number}
}
\description{
The function \code{safecrossover} a graph showing driver conflict of two vehicles changing lanes.
}
\examples{
safecrossover(67, 14)
}
