% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic.R
\name{schematic}
\alias{schematic}
\title{The function \code{schematic} defines key features of a simple bottleneck}
\usage{
schematic()
}
\description{
The function \code{schematic} defines key features of a simple bottleneck
}
\examples{
schematic()
}
