% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedboxplot.R
\name{speedboxplot}
\alias{speedboxplot}
\title{The function \code{speedboxplot} plots the data from a data frame \code{QUKtdayX}}
\usage{
speedboxplot(QUKtdayX)
}
\arguments{
\item{QUKtdayX, }{a matrix}
}
\description{
The function \code{speedboxplot} plots the data from a data frame \code{QUKtdayX}
}
