% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedplot.R
\name{speedplot}
\alias{speedplot}
\title{A scatter plot of traffic data.}
\usage{
speedplot(QUKtdayX)
}
\arguments{
\item{QUKtdayX, }{a matrix}
}
\description{
A scatter plot of traffic data.
}
\examples{
speedplot(QUKtdayX)
}
