\name{speedts}
\alias{speedts}
\docType{data}
\title{\code{speedts} speed data used by \code{speedtsplot}}
\description{
\code{speedts} speed data for lead and following vehicles.}
\usage{data("speedts")}
\format{
  A data frame with  observations on the following 3 variables.
  \describe{
    \item{\code{day}}{a numeric vector}
    \item{\code{t}}{a numeric vector}
    \item{\code{u}}{a numeric vector}
  }
}
\keyword{datasets}
