% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedtsplot.R
\name{speedtsplot}
\alias{speedtsplot}
\title{The function \code{speedtsplot} plots the data from the \code{speedts} data frame.}
\usage{
speedtsplot(speedts)
}
\arguments{
\item{speedts, }{a matrix}
}
\description{
The function \code{speedtsplot} plots the data from the \code{speedts} data frame.
}
