% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0.R
\docType{data}
\name{t0}
\alias{t0}
\title{\code{t0} is the time the analysis is initiated.}
\format{\describe{
\item{t0}{time, measured in seconds.}
}}
\usage{
t0
}
\description{
\code{t0} is the time the analysis is initiated.
}
\keyword{datasets}
