% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1.R
\docType{data}
\name{t1}
\alias{t1}
\title{\code{t1} is time the following vehicle realizes the lead vehicle is decelerating.}
\format{The following vehicle initiates deceleration.
\describe{
\item{t1}{time, measured in seconds.}
}}
\usage{
t1
}
\description{
\code{t1} is time the following vehicle realizes the lead vehicle is decelerating.
}
\keyword{datasets}
