% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2.R
\docType{data}
\name{t2}
\alias{t2}
\title{\code{t2} is time the following vehicle realizes the lead vehicle is decelerating.}
\format{The following vehicle sight-line intersects the lead vehicle trajectory.
\describe{
\item{t2}{time, measured in seconds.}
}}
\usage{
t2
}
\description{
\code{t2} is time the following vehicle realizes the lead vehicle is decelerating.
}
\keyword{datasets}
