% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3.R
\docType{data}
\name{t3}
\alias{t3}
\title{\code{t3} is time the lead vehicle completes deceleration.}
\format{The lead vehicle is traveling at \code{u3} = \code{u0}, the breakdown speed at time \code{t3}.
\describe{
\item{t3}{time, measured in seconds.}
}}
\usage{
t3
}
\description{
\code{t3} is time the lead vehicle completes deceleration.
}
\keyword{datasets}
