% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\docType{data}
\name{theta}
\alias{theta}
\title{The exponential deceleration rate \code{theta} of the lead vehicle.}
\format{\code{theta} is the decelerate from speed \code{upre} to \code{u0} over time \code{t0} to \code{t3}.
\describe{
\item{theta}{lead vehicle deceleration rate, measured in feet per second.}
}}
\usage{
theta
}
\description{
The exponential deceleration rate \code{theta} of the lead vehicle.
}
\keyword{datasets}
