% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryab.R
\name{trajectoryab}
\alias{trajectoryab}
\title{\code{trajectoryab} fits a second-order velocity model and produces a \code{t-x} trajectory
for \code{vehicle = i} where \code{i = 3,4, ..., nveh}.}
\usage{
trajectoryab(tstart, tend, ustart, uend, xstart, xend, step)
}
\arguments{
\item{tstart}{time, a number}

\item{tend}{time, a number}

\item{ustart}{speed, a number}

\item{uend}{speed, a number}

\item{xstart}{location, a number}

\item{xend}{location, a number}

\item{step}{line type, a number}
}
\description{
\code{trajectoryab} fits a second-order velocity model and produces a \code{t-x} trajectory
for \code{vehicle = i} where \code{i = 3,4, ..., nveh}.
}
