% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryab3.R
\name{trajectoryab3}
\alias{trajectoryab3}
\title{\code{trajectoryab3} fits a second-order velocity model and produces a \code{t-x} trajectory
for \code{vehicle = i} where \code{i = 3,4, ..., nveh}.}
\usage{
trajectoryab3(tstart, tend, ustart, uend, xstart, xend, step,type)
}
\arguments{
\item{tstart}{time, a number}

\item{tend}{time, a number}

\item{ustart}{speed, a number}

\item{uend}{speed, a number}

\item{xstart}{location, a number}

\item{xend}{location, a number}

\item{step}{line type, a number}

\item{type}{TRUE to create plots or FALSE otherwise, a logical}
}
\description{
\code{trajectoryab3} fits a second-order velocity model and produces a \code{t-x} trajectory
for \code{vehicle = i} where \code{i = 3,4, ..., nveh}.
}
