% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuxvfix3.R
\name{tuxvfix3}
\alias{tuxvfix3}
\title{\code{tuxvfix3} fixes \code{hsafe} violations for vehicle \code{veh} in matrix \code{tuxv}.}
\usage{
tuxvfix3(i, vehorder, nveh, tuxv, ufix, xfix)
}
\arguments{
\item{i}{index, a number}

\item{vehorder}{vehicle, a vector}

\item{nveh}{number of vehicle in \code{df}, a number}

\item{tuxv}{a data frame created in\code{bk2trials}, a matrix}

\item{ufix}{new speeds, a vector}

\item{xfix}{new locations, a vector}
}
\description{
\code{tuxvfix3} fixes \code{hsafe} violations for vehicle \code{veh} in matrix \code{tuxv}.
}
