% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uab.R
\name{uab}
\alias{uab}
\title{Estimate distance \code{u} using second-order speed function.}
\usage{
uab(u0,a,b,t,t0)
}
\arguments{
\item{u0}{speed at \code{t0}, a number}

\item{a}{second-order speed function parameter, a number}

\item{b}{second-order speed function parameter, a number}

\item{t}{forecast time, a number}

\item{t0}{initial time \code{t0}, a number}
}
\description{
Estimate distance \code{u} using second-order speed function.
}
\examples{
uab(0,16,0,10,0)
}
