% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubreakdown.R
\name{ubreakdown}
\alias{ubreakdown}
\title{The \code{ubreakdown} function generates a breakdown speed vector \code{ubk} assuming a lognormal distribution.}
\usage{
ubreakdown(nveh, ubk, ubk.sd)
}
\arguments{
\item{nveh}{number of vehicles in the investigation or \code{vehicle} = {1, 2, ..., \code{nveh}}, a number}

\item{ubk}{average breakdown speed, a number}

\item{ubk.sd}{standard deviation of \code{ubk}, a number}
}
\description{
The \code{ubreakdown} function generates a breakdown speed vector \code{ubk} assuming a lognormal distribution.
}
