% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf0.R
\docType{data}
\name{uf0}
\alias{uf0}
\title{The following \code{f} vehicle speed \code{uf0} at time \code{t0}.}
\format{The following vehicle speed at \code{t0} is assumed to be independent of the lead vehicle speed \code{ul0}.
\describe{
\item{uf0}{speed, measured in feet per second.}
}}
\usage{
uf0
}
\description{
The following \code{f} vehicle speed \code{uf0} at time \code{t0}.
}
\keyword{datasets}
