% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf1.R
\docType{data}
\name{uf1}
\alias{uf1}
\title{The following \code{f} vehicle speed \code{uf1} at time \code{t1}.}
\format{The following vehicle speed at \code{t1} is assumed to be dependent on the lead vehicle speed \code{ul1}.
\describe{
\item{uf1}{speed, measured in feet per second.}
}}
\usage{
uf1
}
\description{
The following \code{f} vehicle speed \code{uf1} at time \code{t1}.
}
\keyword{datasets}
