% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf2.R
\docType{data}
\name{uf2}
\alias{uf2}
\title{The following \code{f} vehicle speed \code{uf2} at time \code{t2}.}
\format{The following vehicle speed at \code{t2} is assumed to be dependent on the lead vehicle speed \code{ul2}.
\describe{
\item{uf2}{speed, measured in feet per second.}
}}
\usage{
uf2
}
\description{
The following \code{f} vehicle speed \code{uf2} at time \code{t2}.
}
\keyword{datasets}
