% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uf3.R
\docType{data}
\name{uf3}
\alias{uf3}
\title{The following \code{f} vehicle speed \code{uf3} at time \code{t0}.}
\format{The following vehicle speed at \code{t3} is assumed to be equal to the breakdown speed \code{u0}.
\describe{
\item{uf3}{speed, measured in feet per second.}
}}
\usage{
uf3
}
\description{
The following \code{f} vehicle speed \code{uf3} at time \code{t0}.
}
\keyword{datasets}
