% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_sdk.R
\name{uk_sdk}
\alias{uk_sdk}
\title{A speed volatility model.}
\usage{
uk_sdk(QUKtdayX, type)
}
\arguments{
\item{QUKtdayX, }{a matrix}

\item{type}{logical}
}
\value{
If \code{type} is TRUE, plot a density-speed plot is created and a data frame is returned.
If If \code{type} is FALSE,  a data frame is returned.
}
\description{
A speed volatility model.
}
\examples{
\donttest{uk_sdk(QUKtdayX, TRUE)}
}
