% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul0.R
\docType{data}
\name{ul0}
\alias{ul0}
\title{The lead \code{l} vehicle speed \code{ul0} at time \code{t0}.}
\format{The lead vehicle speed at \code{t0} is assumed to be independent of the pre-breakdown speed \code{upre}.
\describe{
\item{ul0}{speed, measured in feet per second.}
}}
\usage{
ul0
}
\description{
The lead \code{l} vehicle speed \code{ul0} at time \code{t0}.
}
\keyword{datasets}
