% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul1.R
\docType{data}
\name{ul1}
\alias{ul1}
\title{The following \code{l} vehicle speed \code{ul1} at time \code{t1}.}
\format{The lead vehicle speed at \code{t1} is estimated with the exponential speed function \code{gbm(upre, theta, t1)}.
\describe{
\item{ul1}{speed, measured in feet per second.}
}}
\usage{
ul1
}
\description{
The following \code{l} vehicle speed \code{ul1} at time \code{t1}.
}
\keyword{datasets}
