% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul2.R
\docType{data}
\name{ul2}
\alias{ul2}
\title{The following \code{l} vehicle speed \code{ul2} at time \code{t2}.}
\format{The lead vehicle speed at \code{t2} is estimated with the exponential speed function \code{gbm(upre, theta, t2)}.
\describe{
\item{ul2}{speed, measured in feet per second.}
}}
\usage{
ul2
}
\description{
The following \code{l} vehicle speed \code{ul2} at time \code{t2}.
}
\keyword{datasets}
