% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul3.R
\docType{data}
\name{ul3}
\alias{ul3}
\title{The following \code{l} vehicle speed \code{ul3} at time \code{t3}.}
\format{The lead vehicle speed at \code{t2} is estimated with the exponential speed function \code{gbm(upre, theta, t3)}.
\describe{
\item{ul3}{speed, measured in feet per second.}
}}
\usage{
ul3
}
\description{
The following \code{l} vehicle speed \code{ul3} at time \code{t3}.
}
\keyword{datasets}
