% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdzipper.R
\name{usdzipper}
\alias{usdzipper}
\title{The \code{usdzipper} function is used to perform a sensitivity analysis of a single vehicles
in traffic breakdown.}
\usage{
usdzipper(tstart, tend, umn, usd, xstart, xend)
}
\arguments{
\item{tstart}{start time, a number}

\item{tend}{end time, a number}

\item{umn}{start speed (mph) for vehicle in lane 1, a number}

\item{usd}{speed volatility for \code{umn}, a number}

\item{xstart}{start location for vehicle in lane 1 (feet), a number}

\item{xend}{end location for vehicle in lane 1 (feet), a number}
}
\value{
\code{usdzipper} uses a deterministic model to illustrate an ``idealistic'' situation,
a so-called a ``zipper merge.''
}
\description{
The \code{usdzipper} function is used to perform a sensitivity analysis of a single vehicles
in traffic breakdown.
}
\examples{
usdzipper(0, 5, 41, 11.6, 0, 500)
}
