% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehdf.R
\name{vehdf}
\alias{vehdf}
\title{\code{vehdf} selects information about vehicle \code{veh}
of the \code{df} matrix for analysis with the \code{brktrials2} function.}
\usage{
vehdf(veh, nveh, df)
}
\arguments{
\item{veh}{vehicle, a number}

\item{nveh}{number of vehicle in \code{df}, a number}

\item{df}{a data frame created in\code{bk2trials}, a data frame}
}
\description{
\code{vehdf} selects information about vehicle \code{veh}
of the \code{df} matrix for analysis with the \code{brktrials2} function.
}
