% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehfix.R
\name{vehfix}
\alias{vehfix}
\title{\code{vehfix} fixes \code{hsafe} violation with speed \code{ufix} and location \code{xfix} and \code{yfix} vectors from \code{brktrials2}}
\usage{
vehfix(veh, nveh, ufix, xfix, yfix, df)
}
\arguments{
\item{veh}{vehicle, a number}

\item{nveh}{number of vehicle in \code{df}, a number}

\item{ufix}{speed, a vector}

\item{xfix}{location, a vector}

\item{yfix}{location, a vector}

\item{df}{a data frame created in \code{bk2trials}, a data frame}
}
\description{
\code{vehfix} fixes \code{hsafe} violation with speed \code{ufix} and location \code{xfix} and \code{yfix} vectors from \code{brktrials2}
}
