% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehid.R
\name{vehid}
\alias{vehid}
\title{\code{vehid} selects information about vehicle \code{veh}
from the \code{tuxv} matrix for analysis with the \code{brktrials3} function.}
\usage{
vehid(veh, tuxv)
}
\arguments{
\item{veh}{vehicle, a number}

\item{tuxv}{a data frame created in\code{brktrials3}, a matrix}
}
\description{
\code{vehid} selects information about vehicle \code{veh}
from the \code{tuxv} matrix for analysis with the \code{brktrials3} function.
}
