% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehtfix.R
\name{vehtfix}
\alias{vehtfix}
\title{\code{vehtfix} fixes \code{hsafe} violations for vehicle \code{veh}.}
\usage{
vehtfix(veh, nveh, df, tstart, tend, step, ufix, xfix)
}
\arguments{
\item{veh}{vehicle, a number}

\item{nveh}{number of vehicle in \code{df}, a number}

\item{df}{a data frame created in\code{bk2trials}, a data frame}

\item{tstart}{time, a number}

\item{tend}{time, a number}

\item{step}{size in seconds, a number}

\item{ufix}{new speeds, a vector}

\item{xfix}{new locations, a vector}
}
\description{
\code{vehtfix} fixes \code{hsafe} violations for vehicle \code{veh}.
}
