% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xab.R
\name{xab}
\alias{xab}
\title{Estimate distance \code{x} using second-order speed function.}
\usage{
xab(x0,u0,a,b,t,t0)
}
\arguments{
\item{x0}{location at \code{t0}, a number}

\item{u0}{speed at \code{t0}, a number}

\item{a}{second-order speed function parameter, a number}

\item{b}{second-order speed function parameter, a number}

\item{t}{forecast time, a number}

\item{t0}{initial time \code{t0}, a number}
}
\description{
Estimate distance \code{x} using second-order speed function.
}
\examples{
xab(0,0,16,0,10,0)
}
