% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xabmerge3.R
\name{xabmerge3}
\alias{xabmerge3}
\title{\code{xabmerge3} produces \code{t-x} trajectories for a pair of lead and following vehicles.}
\usage{
xabmerge3(df1, df2, leff, step, k, type)
}
\arguments{
\item{df1}{leading vehicle, a matrix}

\item{df2}{following vehicle, a matrix}

\item{leff}{vehicle length, a number}

\item{step}{time-step size, a number}

\item{k}{counter, a number}

\item{type}{logical for plotting where TRUE creates a plot}
}
\value{
The \code{xabmerge3}, a wrapper function for \code{bmfree2}, \code{xabparam} and
\code{hsafe}, returns a smooth \code{hsafe} rule \code{t-x} trajectory.
for the following vehicle. The lead vehicle trajectory is not affected.
Incidentally, inputs \code{df1} and \code{df2} were created with  \code{bmfree2} and used
for testing.
}
\description{
\code{xabmerge3} produces \code{t-x} trajectories for a pair of lead and following vehicles.
}
