% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xabparam.R
\name{xabparam}
\alias{xabparam}
\title{Estimate the parameters \code{a} and \code{b} of a first-order acceleration model \code{a + b * t}.}
\usage{
xabparam(tstart, tend, ustart, uend, xstart, xend)
}
\arguments{
\item{tstart}{time in seconds, a number}

\item{tend}{time in seconds, a number}

\item{ustart}{speed, a number}

\item{uend}{speed, a number}

\item{xstart}{location, a number}

\item{xend}{location, a number}
}
\description{
Estimate the parameters \code{a} and \code{b} of a first-order acceleration model \code{a + b * t}.
}
\examples{
xabparam(0, 18.59797, 92.4, 0, 0, 1397.045)
}
