% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf0.R
\docType{data}
\name{xf0}
\alias{xf0}
\title{The lead \code{f} vehicle location \code{xf0} at time \code{t0}.}
\format{The following vehicle location at \code{t0} is assumed to be dependent space headway \code{h0}.
\describe{
\item{xf0}{distance, measured in feet.}
}}
\usage{
xf0
}
\description{
The lead \code{f} vehicle location \code{xf0} at time \code{t0}.
}
\keyword{datasets}
