% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf1.R
\docType{data}
\name{xf1}
\alias{xf1}
\title{The lead \code{f} vehicle location \code{xf1} at time \code{t1}.}
\format{The following vehicle location at \code{t1} is assumed to be dependent space headway \code{h1}.
\describe{
\item{xf1}{distance, measured in feet.}
}}
\usage{
xf1
}
\description{
The lead \code{f} vehicle location \code{xf1} at time \code{t1}.
}
\keyword{datasets}
