% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf2.R
\docType{data}
\name{xf2}
\alias{xf2}
\title{The lead \code{f} vehicle location \code{xf2} at time \code{t2}.}
\format{The following vehicle location at \code{t2} is assumed to be dependent space headway \code{h2}.
\describe{
\item{xf2}{distance, measured in feet.}
}}
\usage{
xf2
}
\description{
The lead \code{f} vehicle location \code{xf2} at time \code{t2}.
}
\keyword{datasets}
