% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xf3.R
\docType{data}
\name{xf3}
\alias{xf3}
\title{The lead \code{f} vehicle location \code{xf3} at time \code{t3}.}
\format{The following vehicle location at \code{t3} is assumed to be dependent space headway \code{h3}.
\describe{
\item{xf3}{distance, measured in feet.}
}}
\usage{
xf3
}
\description{
The lead \code{f} vehicle location \code{xf3} at time \code{t3}.
}
\keyword{datasets}
