% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xfollow.R
\name{xfollow}
\alias{xfollow}
\title{Estimate \code{x} distance traveled using a constant speed model.}
\usage{
xfollow(x0, u, t, t0)
}
\arguments{
\item{x0}{starting location at time \code{t0}, a number}

\item{u}{speed, a number}

\item{t}{travel time, a number}

\item{t0}{deceleration time, a number}
}
\description{
Estimate \code{x} distance traveled using a constant speed model.
}
