\name{xl0}
\alias{xl0}
\docType{data}
\title{The lead \code{l} vehicle location \code{xl0} at time \code{t0}.}
\format{The following vehicle location at \code{t0} is assumed to be dependent space headway \code{h0}.
\describe{
\item{xl0}{distance, measured in feet.}
}}
\usage{
xl0
}
\description{
The lead \code{l} vehicle location \code{xl0} at time \code{t0}.
}
\keyword{datasets}
