\name{xl1}
\alias{xl1}
\docType{data}
\title{The lead \code{l} vehicle location \code{xl1} at time \code{t1}.}
\format{The lead vehicle location at \code{t1} is assumed to be dependent space headway \code{h1}.
\describe{
\item{xl1}{distance, measured in feet.}
}}
\usage{
xl1
}
\description{
The lead \code{l} vehicle location \code{x1} at time \code{t1}.
}
\keyword{datasets}
