\name{xl2}
\alias{xl2}
\docType{data}
\title{The lead \code{l} vehicle location \code{xl2} at time \code{t2}.}
\format{The lead vehicle location at \code{t2} is assumed to be dependent space headway \code{h2}.
\describe{
\item{xf2}{distance, measured in feet.}
}}
\usage{
xl2
}
\description{
The lead \code{l} vehicle location \code{xl2} at time \code{t2}.
}
\keyword{datasets}
