\name{xl3}
\alias{xl3}
\docType{data}
\title{The lead \code{l} vehicle location \code{xl3} at time \code{t3}.}
\format{The lead vehicle location at \code{t3} is assumed to be dependent space headway \code{h3}.
\describe{
\item{xl3}{distance, measured in feet.}
}}
\usage{
x3
}
\description{
The lead \code{l} vehicle location \code{xl3} at time \code{t3}.
}
\keyword{datasets}
