\name{xstart}
\alias{xstart}
\docType{data}
\title{
\code{xstart} is a vector that designates the start \code{x} location for the \code{brktrials} function.
}
\format{A vector used by \code{brktrials}
\describe{
\item{xstart}{\code{xstart} is measured in feet}
}}
\description{ \code{xstart} is a input vector to \code{brtrials}. For example, the start locations of the vehicles may be: -900 -990 -1000 -1090 -1100 feet.}
\keyword{datasets}
